requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.2
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
streamlit==1.31.1
audio_recorder_streamlit==0.0.8
python-dotenv==1.0.1
httpx<0.28.0
neo4j==5.20.0
spacy==3.7.4
faiss-cpu==1.7.4
pyvis==0.3.2
groq==0.11.0
aiofiles==23.2.1
python-multipart==0.0.9
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
apscheduler==3.10.4
elevenlabs==1.2.2